refactor: Header 컴포넌트의 삼항 연산자를 이용한 조건부 렌더링 최소화
